fix: remediate dependency advisories
