+++ package.json
@@ -19,7 +19,7 @@
     "react-datepicker": "^9.1.0",
     "react-dom": "^19.0.0",
     "recharts": "^3.8.1",
-    "xlsx": "^0.18.5"
+    "write-excel-file": "^4.1.1"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
@@ -29,8 +29,5 @@
   },
   "engines": {
     "node": ">=24"
-  },
-  "pnpm": {
-    "ignoredBuiltDependencies": ["sharp", "unrs-resolver"]
   }
 }
