+++ pnpm-workspace.yaml
@@ -1,3 +1,9 @@
+overrides:
+  '@babel/core': 7.29.6
+  form-data: 4.0.6
+  js-yaml: 4.2.0
+  postcss: 8.5.16
+
 allowBuilds:
   sharp: true
   unrs-resolver: true
