+++ src/app/components/exchange-rate-export-files.js
@@ -1,0 +1,20 @@
+export function rowsToCsv(rows) {
+  return `${rows.map((row) => row.map(formatCsvCell).join(",")).join("\r\n")}\r\n`;
+}
+
+export function downloadTextFile(text, filename, type) {
+  const blob = new Blob([text], { type });
+  const url = URL.createObjectURL(blob);
+  const link = document.createElement("a");
+  link.href = url;
+  link.download = filename;
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+  URL.revokeObjectURL(url);
+}
+
+function formatCsvCell(value) {
+  const text = value == null ? "" : String(value);
+  return /[",\r\n]/.test(text) ? `"${text.replaceAll("\"", "\"\"")}"` : text;
+}
